Formatter tests
